Adding a collapsing menu to choose the coloration for the langage
I added a way to change which language to use for syntax highlighting.

For this, I added a function in editor.js that take the mode to use. I added a message to be sent when the menu is changed.
And finally I put the menu.

The next step is to catch a parameter in the get payload to allow the user to select a mode to be used directly from the url.