The language of the Ace Editor can now be set through the "editorLanguage" get parameter
The editor now react when you put a paramater in the get. If the language is known in the list set in frontend/common/aceMode.js, the editor will be set to this one. If not, it will stick to C++.

I will now try to set up a bit better the language menu and make that changing the language also set the current url without reloading the page.
The pull request will have the menu commented out so you can use it if you wish to.